gym: train the navigator each cycle on the live space (overnight engine)
GenesisEvalAppRuntime.TrainNavigatorCycle (new partial, gated via the single _modelOpsGate): samples live members, derives the cue from the LIVE relation depth (ClimbAncestors by increasing relational degree -> GENUS/DOMAIN/ROOT; flat space = GENUS only), builds flow-field query trajectories, runs a few Adam epochs on the shared State.Navigator, probes resolve%, folds traversed concepts into the engine self (PerceiveSelf -> the vital loop). Hooked into GymLoopAsync.onCycle every cycle (bounded, sub-second on GPU; try/catch never stops the gym), logged + stashed (LastNavTrain/NavResolveRunningPct). Checkbox CurTrainNavigator default ON. f32 navigator weights (~564->282MB/autosave; legacy f64 shard still decodes; round-trip bit-exact). GPU-OOM-safe (CPU fallback, net rests on CPU between cycles). Test: loss 0.287->0.000, resolve 0->67% over 6 cycles, gated. Full suite 250/0/45.

+++ src/GenesisNova/Cognition/Navigator/NavQueryPolicyNet.cs
@@ -160,16 +160,17 @@ public sealed class NavQueryPolicyNet : Module
     /// <summary>Total trainable parameter count (diagnostic — proves the net stays a thin controller).</summary>
     public long ParameterCount() => parameters().Sum(p => p.numel());
 
-    /// <summary>EXPORT every named weight tensor for the checkpoint (as f64 — lossless from the net's f32). The
-    /// architecture dims ride along so a load can detect a re-architected navigator and decline rather than crash.</summary>
+    /// <summary>EXPORT every named weight tensor for the checkpoint as NATIVE f32 (the net's own dtype — lossless, and
+    /// HALF the bytes of the former f64 export, so autosave I/O is halved). The architecture dims ride along so a load
+    /// can detect a re-architected navigator and decline rather than crash.</summary>
     public NavigatorSnapshot ExportWeights()
     {
         using var _ = no_grad();
         var list = new List<NavParameterSnapshot>();
         foreach (var (name, p) in named_parameters())
         {
-            using var t = p.detach().to(float64).cpu().contiguous();
-            list.Add(new NavParameterSnapshot(name, p.shape.ToArray(), t.data<double>().ToArray()));
+            using var t = p.detach().to(float32).cpu().contiguous();
+            list.Add(new NavParameterSnapshot(name, p.shape.ToArray(), t.data<float>().ToArray()));
         }
         return new NavigatorSnapshot(Dim, Hidden, CueCount, SelfLength, list.ToArray());
     }
