+++ src/GenesisNova/Persistence/GenesisCheckpoint.cs
@@ -54,8 +54,8 @@ public sealed record GenesisCheckpoint(
     GrammarRoleSnapshot[]? GrammarRoles = null,
     // THE PLATONIC NAVIGATOR policy net (NavQueryPolicyNet) weights — the trained query-conditioned controller that
     // walks the substrate. Null on pre-navigator checkpoints (then the runtime keeps its freshly-initialised navigator).
-    // The big parameter tensors are stored as a SEPARATE f64 shard (see GenesisShardedCheckpointStore), like the model
-    // NN — so an overnight-trained navigator survives an app restart. Persisted so the navigator isn't reset to
+    // The big parameter tensors are stored as a SEPARATE native-f32 shard (see GenesisShardedCheckpointStore), like the
+    // model NN — so an overnight-trained navigator survives an app restart. Persisted so the navigator isn't reset to
     // untrained on every load (the same drop-on-load gap previously fixed for the GRU/route/role heads).
     NavigatorSnapshot? Navigator = null,
     // THE PERSISTENT SELF (GenesisInferenceEngine._selfField) — the decaying meaning-space standing wave the mind
@@ -74,13 +74,14 @@ public sealed record GenesisCheckpoint(
 public sealed record GrammarRoleSnapshot(string Token, int Present, int Absent, int AsAnswer, int AsCopula);
 
 /// <summary>Serializable weights of the navigator policy net (<see cref="GenesisNova.Cognition.Navigator.NavQueryPolicyNet"/>).
-/// <see cref="Parameters"/> are the net's named tensors (flattened, row-major) — concatenated into one f64 shard on save;
-/// the architecture dims let load reject a shape-mismatched (re-architected) navigator and stay fresh rather than throw.</summary>
+/// <see cref="Parameters"/> are the net's named tensors (flattened, row-major) — concatenated into one native-f32 shard on
+/// save (half the bytes of the former f64 encoding, lossless for the f32 params); the architecture dims let load reject a
+/// shape-mismatched (re-architected) navigator and stay fresh rather than throw.</summary>
 public sealed record NavigatorSnapshot(int Dim, int Hidden, int CueCount, int SelfLength, NavParameterSnapshot[] Parameters);
 
 /// <summary>One named navigator parameter tensor: its module path <see cref="Name"/>, its <see cref="Shape"/>, and its
-/// row-major <see cref="Values"/> (emptied in the manifest's meta JSON — the bytes live in the concatenated f64 shard).</summary>
-public sealed record NavParameterSnapshot(string Name, long[] Shape, double[] Values);
+/// row-major f32 <see cref="Values"/> (emptied in the manifest's meta JSON — the bytes live in the concatenated f32 shard).</summary>
+public sealed record NavParameterSnapshot(string Name, long[] Shape, float[] Values);
 
 public sealed record MatrixSnapshot(int Rows, int Cols, double[] Values)
 {
